GTFS transform

+++ src/main/java/com/conveyal/datatools/editor/jobs/ExportSnapshotToGTFSJob.java
@@ -16,16 +16,21 @@ import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
 
+/**
+ * This job will export a database snapshot (i.e., namespace) to a GTFS file. If a feed version is supplied in the
+ * constructor, it will assume that the GTFS file is intended for ingestion into Data Tools as a new feed version.
+ */
 public class ExportSnapshotToGTFSJob extends MonitorableJob {
 
     private static final Logger LOG = LoggerFactory.getLogger(ExportSnapshotToGTFSJob.class);
     private final Snapshot snapshot;
-    private final String feedVersionId;
+    private final FeedVersion feedVersion;
+    private File tempFile;
 
-    public ExportSnapshotToGTFSJob(Auth0UserProfile owner, Snapshot snapshot, String feedVersionId) {
+    public ExportSnapshotToGTFSJob(Auth0UserProfile owner, Snapshot snapshot, FeedVersion feedVersion) {
         super(owner, "Exporting snapshot " + snapshot.name, JobType.EXPORT_SNAPSHOT_TO_GTFS);
         this.snapshot = snapshot;
-        this.feedVersionId = feedVersionId;
+        this.feedVersion = feedVersion;
         status.update("Starting database snapshot...", 10);
     }
 
@@ -40,7 +45,9 @@ public class ExportSnapshotToGTFSJob extends MonitorableJob {
 
     @Override
     public void jobLogic() {
-        File tempFile;
+        // Determine if storing/publishing new feed version for snapshot. If not, all we're doing is writing the
+        // snapshot to a GTFS file.
+        boolean isNewVersion = feedVersion != null;
         try {
             tempFile = File.createTempFile("snapshot", "zip");
         } catch (IOException e) {
@@ -52,13 +59,14 @@ public class ExportSnapshotToGTFSJob extends MonitorableJob {
         FeedLoadResult result = exporter.exportTables();
         if (result.fatalException != null) {
             status.fail(String.format("Error (%s) encountered while exporting database tables.", result.fatalException));
+            return;
         }
 
         // Override snapshot ID if exporting feed for use as new feed version.
-        String filename = feedVersionId != null ? feedVersionId : snapshot.id + ".zip";
-        String bucketPrefix = feedVersionId != null ? "gtfs" : "snapshots";
+        String filename = isNewVersion ? feedVersion.id : snapshot.id + ".zip";
+        String bucketPrefix = isNewVersion ? "gtfs" : "snapshots";
         // FIXME: replace with use of refactored FeedStore.
-        // Store the project merged zip locally or on s3
+        // Store the GTFS zip locally or on s3.
         status.update("Writing snapshot to GTFS file", 90);
         if (DataManager.useS3) {
             String s3Key = String.format("%s/%s", bucketPrefix, filename);
@@ -66,17 +74,24 @@ public class ExportSnapshotToGTFSJob extends MonitorableJob {
             LOG.info("Storing snapshot GTFS at s3://{}/{}", DataManager.feedBucket, s3Key);
         } else {
             try {
-                FeedVersion.feedStore.newFeed(filename, new FileInputStream(tempFile), null);
+                File gtfsFile = FeedVersion.feedStore.newFeed(filename, new FileInputStream(tempFile), null);
+                if (isNewVersion) feedVersion.assignGtfsFileAttributes(gtfsFile);
             } catch (IOException e) {
                 status.fail(String.format("Could not store feed for snapshot %s", snapshot.id), e);
             }
         }
-        // Delete snapshot temp file.
-        tempFile.delete();
     }
 
     @Override
     public void jobFinished () {
         if (!status.error) status.completeSuccessfully("Export complete!");
+        // Delete snapshot temp file.
+        if (tempFile != null) {
+            LOG.info("Deleting temporary GTFS file for exported snapshot at {}", tempFile.getAbsolutePath());
+            boolean deleted = tempFile.delete();
+            if (!deleted) {
+                LOG.warn("Temp file {} not deleted. This may contribute to storage space shortages.", tempFile.getAbsolutePath());
+            }
+        }
     }
 }
